5.0

+++ src/ru/job4j/converter/Converter.java
@@ -5,14 +5,14 @@ public class Converter {
         int rsl = value / 70;
         return rsl;
     }
-    public static int rubleToDollar(int value){
-        int rsl = value / 60;
+    public static double rubleToDollar(double value){
+        double rsl = value / 60;
         return rsl;
     }
     public static void main(String[] args){
         int euro = Converter.rubleToEuro(140);
         System.out.println("140 rubles are " + euro + " euro.");
-        int dollar = Converter.rubleToDollar(140);
+        double dollar = Converter.rubleToDollar(140);
         System.out.println("140 rubles are " + dollar + " dollar.");
     }
 }
